build: update to ZTL 0.25.1

+++ include/z21/loco_info.hpp
@@ -85,9 +85,10 @@ constexpr int32_t decode_rvvvvvvv(LocoInfo::SpeedSteps speed_steps,
 constexpr uint8_t
 encode_rvvvvvvv(LocoInfo::SpeedSteps speed_steps, bool dir, int32_t speed) {
   // Halt
-  if (!speed) return static_cast<uint8_t>(dir << 7u);
+  if (!speed) return static_cast<uint8_t>(static_cast<uint32_t>(dir) << 7u);
   // EStop
-  else if (speed < 0) return static_cast<uint8_t>(dir << 7u) | 0b1u;
+  else if (speed < 0)
+    return static_cast<uint8_t>(static_cast<uint32_t>(dir) << 7u | 0b1u);
 
   auto vvvvvvv{static_cast<uint32_t>(speed + 1)};
 
@@ -97,7 +98,7 @@ encode_rvvvvvvv(LocoInfo::SpeedSteps speed_steps, bool dir, int32_t speed) {
     if (!(speed % 2)) vvvvvvv |= ztl::mask<4u>;
   }
 
-  return static_cast<uint8_t>(static_cast<uint32_t>(dir << 7u) | vvvvvvv);
+  return static_cast<uint8_t>(static_cast<uint32_t>(dir) << 7u | vvvvvvv);
 }
 
 } // namespace z21
